backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.1
python-multipart==0.0.9
chromadb==0.5.23
google-generativeai==0.8.3
